lab4/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
pillow==10.2.0
numpy==1.26.4
python-jose==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.3
celery==5.3.6
pydantic[email]==2.6.4

